requirement.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
google-generativeai>=0.3.2
google-auth>=2.25.2
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
gspread>=5.12.4
pandas>=2.2.0
plotly>=5.17.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
httpx>=0.25.2
pytest>=7.4.3
pytest-asyncio==0.21.1
kaleido==0.2.1
